Reviewed as ENG-307 by will, changes as follows:
The EventBase hierarcy of objects are used to communicate happenings within the METC framework, e.g. incoming market data.  Some of the objects include a symbol attribute.  Heretofore, the type of this 
attribute was String.  This change makes that type MSymbol, which is correct in the agnostic lingua-franca.  This change is needed to properly implement depth-of-book requests for the Bogus feed.  The 
EventBase family of objects have minimal unit tests.  A following change related to the Bogus feed changes includes more extensive tests.

The most significant functional implication of this change is that existing esper queries acting on the aforementioned symbol attribute *must* change as follows:

"select * from trade where symbol='METC'"
-to-
"select * from trade where symbolAsString='METC'"

This change has been approved by PS (and engineering, for that matter).  Complimentary with this review are a JIRA issue to add this change to release notes and a change on the private side.  I ran all 
unit tests successfully (mvn cmdline and Photon).